Rename internal/test to psltest
